test(raft): add a randomized liveness test
This is similar to the correctness test that already exists. First we execute a random sequence of operation, which may contain "failures" like messages are lost or delayed. Then we execute a fixed sequence of operations without any "failures". The idea is that if there are no message lost, then eventually a leader should be elected, and eventually all entries are replicated to all replicas, and all entries are committed.