fix(raft): commit entries when response for heartbeats are received
If a previous ack was lost, and the followers have already received the latest logs, then the leader will send only heartbeats. In this case, previously, the leader did not commit because it did not update the commit index when an ack for heartbeat is received. This can cause problems because as long as new entries are not written the existing last few entries will never get committed.

(cherry picked from commit 3e4556c42ce46f63d85e91666020b57e99a06af4)