test(clients/go): add debug logging
It can happen that builds fail due to invalid JSON in the tests, most
likely because the response is not a valid response but some error.
Unfortunately it's hard to debug because we can't see the actual invalid
output. This commit simply logs that as well on error.

(cherry picked from commit 05b1d1c1cf0fb38d395c08f8e63cd58b27fbf375)