feat(expression-language): add getListOfStrings to evaluation results
Adds a new method to evaluation result that allows to retrieve a list of
strings. In the same fashion as other methods, it returns null if that
could not be achieved, i.e. it returns null if it is not an array or if
any of the items is not a string.