feat(engine): dont allow non-strings in array of strings expr
Previously, as a quick solution, we simply converted each of the items
using our knowledge that the items are formatted as MsgPack objects.
However, we should not build upon that knowledge, since it's an
implementation detail of the feel engine.

Instead, the evaluation result now provides a new getListOfStrings
method that returns null if the expression did not evaluate to an array,
and also returns null if the expression evaluated to an array where at
least one of the items is not a String.

This switches the usage and builds a failure message specifically for
this case.