test(atomix/cluster): fixes flaky priority election test
The test currently uses two boolean flags which represent that the
election was toggled, which it spies on. At the end, it then asserts
that the `set` calls for each were called in the right order.

This can break since the timers renew themselves automatically, meaning
that once the high prio timer has expired after 100ms, it will
automatically renew itself, such that its trigger will be called after
100ms. It's possible then that it is called again before the low
priority timer is called. The low priority timer has a target priority
of 4, and an initial priority of 1. This means it has to do 3 rounds, so
roughly 300ms, before it will actually trigger. This gives ample time to
the high priority timer to trigger more than once.

The test is here rewritten to avoid relying on a strict ordering of
method calls, and instead just checks that:

- the first timer called is the high priority timer
- the low priority timer is eventually called after (with any number of
  high priority calls in between)

(cherry picked from commit a8a720598b9b74d738e78903d95639151bcf70ff)