feat(engine): restrict allowed task headers
Usage of the special defined task header should not be allowed as these
are used by the engine.

At this time, the engine uses the special assignee, candidateGroups and
formKey headers (current all prefixed with `io.camunda.zeebe:`) to store
specialized values in order to communicate them to tools outside of
Zeebe (e.g. Tasklist). Some of these headers add additional restrictions
on what the value is allowed to be. Allowing users to set their own
values can produce problems (albeit missing candidate groups for a task
in tasklist) where the user might not understand what they did wrong.

By restricting user defined processes to not be allowed to use the
restricted task headers, it pushes a user to look up the documentation
to understand how to use that specific feature.

For example, when a user tries to define the candidate groups property
as a task header, their process will be rejected during deployment. The
error message will tell them that the specific key is not allowed to be
the restricted header 'io.camunda.zeebe:candidateGroups'. The user will
then question: but how should I define my candidate groups then? And we
expect the user to look up the documentation at this point. The docs
will then inform them to use the specialized TaskAssignment's
candidategroups attribute.