ci: validate benchmark name
The create benchmark script uses the benchmark name among others for the
k8s namespace. However, k8s does not accept dots in the namespace. I
found this out after fully building Zeebe and the docker images locally,
which takes a few minutes. We can simply check for this at the start of
the script to fail early and help users of the script with such
mistakes.