doc: refer to bors.toml in os-smoke-test
Since bors now waits for the successful completion of the os-smoke-test jobs, we should remember to update the statuses when we change this name, or more likely, change the matrix. This is because the bors.toml refers to the actual job name produced from the name template and the matrix.