build(ci): fix junit report pattern
Fixes the Jenkins junit plugin pattern used to detect test reports in
our CI pipeline. The flaky test plugin will add a `-flaky` suffix, so we
need to add an additional `*` to the pattern. Unfortunately as it's not
a regex but simply an Ant glob pattern, there is no way to further
restrict matches, so there's still the risk it could collide and double
the test reports.