build: add recreateBenchmark.sh script
This script allows you to easily recreate an existing benchmark that was
created with createBenchmark.sh. It simply builds the docker images and
updates the deployment.

An additional check that the k8s namespace is available is added to
this, as a quick check that you're using the correct namespace and are
on the correct context. Or perhaps the user was actually intending to
run ./createBenchmark.sh instead.

Sadly, the make update does not restart the pods. We should consider
solutions for that. However, as soon as the pods restart, the pods are
updated because the pullPolicy is set to always. Therefore this applies
a hard delete of all the broker and gateway pods to force the update.
This is not a rolling update.