fix(createMedicCWBenchmark): use ls -t1 to find the oldest benchmark
Using the name to figure out the latest benchmark fails in the new year,
because cw-01 is before cw-51. We can use ls -t1 to sort the output of
ls on timestamp and make sure only the name is returned. We then don't
need to sort again. This ls command outputs the files in chronological
order, so we'll need to swap which specific name we choose from head to
tail and from tail to head.