fix: print correct json input
Previously on test failure, the input was overwritten by the return of the reformatJson function, which gave not much helpful error message. The return value is stored in a intermediate variable to be overwrite the real variable only in the case of success, such that we print the correct input.

(cherry picked from commit 0913760cf597328498f78efd9dd871733164993a)