docs(CONTRIBUTING): fix markdown formatting
This line used a `#` character on a new line without being escaped. That
means it will be counted as an ATX heading. To avoid that, we can either
escape the character or move it into the paragraph line. However,
escaping the character leads to unexpected behavior in other flavors
like on GitHub. It would show the code span of the backtick string with
the escape character visibile as it is not regarded an escape character
in that position.