docs(.github): fix backport_template markdown formatting
ATX heading marker characters (i.e. `#`), need to be escaped in order to
not be counted as a new heading.

Strangely, spotless-flexmark was taking the `closes #` line and replaced
it with `close-->`, but I'm not sure why. Adding an additional non-empty
line under it solves the issue. I'll probably open an issue about this
in flexmark-java.