docs(CONTRIBUTING): fix markdown blockquotes formatting
According to the commonmark specs, 2 blockquotes can only be seperate
from each other with a blank line in between them. Otherwise it is
counted as 1 blockquote of multiple paragraphs. Strangely,
spotless-flexmark seems to disregard this rule entirely. I'll open an
issue for it in flexmark-java. For now, we'll have to live with a single
blockquote of 3 paragraphs here, which in my opinion doesn't really
affect the understandability much.