fix: ignore benchmark deletion before CW 5
We have currently not the optimal logic to delete benchmarks before calendar week 5, since our naming schema contains only the calendar weeks. This means if we are in calendar week 2 we delete 01 and not 49. This commit turns off deletion for the begin of the year and print a notification to delete it manually.