build(gateway-protocol-impl): mark gateway-protocol as compile time dependency
Marks the gateway-protocol module as a compile dependency only to avoid
licensing issues. The protocol is licensed under the Zeebe Community
License, whereas the implementation module is licensed under Apache 2.
As this is meant to be embedded into other applications, it's better to
use the more permissive and understood of the two licenses.

Additionally, the protocol dependency is really only there to ensure
that the implementation module always runs after during a maven build,
allowing the build to pre-process or generate any resources for the
protocol module before the implementation's.