feat(dmn): add eval decision api
Adds the interface method to expose decision evaluation. Originally,
this was thought to take a ParsedDecision, but it became clear this
should take the entire parsed DRG and simply refer to the specific
decision to evaluate by its id.

The naming DecisionResult and DecisionContext also differ from the
original EvaluationResult and EvaluationContext, because this allows us
to more easily differentiate from the expression language module, which
already uses these terms.