fix: await exporter position update
Previously we awaited that the director has seen all records and expected that the position is already updated. This can't be guaranteed. The position was updated after the collection, which lead to race conditions. Furthermore the collection was previously not thread safe. In general it is not necessary to await that the record is read, it is enough to await the position update. This commit fixes this test behavior.

(cherry picked from commit dac22c6f382efd261c11fc41c2c1de63ac51e36c)