refactor(engine): use event sourcing terminology
Since event sourcing was introduced the processing state machine
generally doesn't process events any more. An exception is that it reads
all records from the log before determining whether that record needs to
be processed, and so it deals with events as well. But generally, it
doesn't process events any more. The naming used in the code should
reflect the objects that it refers to.

So: we read records, and process commands. And at the end of processing
a command, we write some records again.