test(engine): synchronously pause the processor
Some tests need to pause the stream processor, but this is an
async method. In the test this would then require the test to check that
the processor was actually paused before continuing its test. We can
more easily achieve this by making the stream processor pausing a
blocking call when used in tests.