test(engine): synchronously pause the processor
Some tests need to pause the stream processor, but this is an
async method. In the test this would then require the test to check that
the processor was actually paused before continuing its test. We can
more easily achieve this by making the stream processor pausing a
blocking call when used in tests.

(cherry picked from commit 026822fa010870adad3a5c3088b7db5a25ee3709)