fix(engine): consider all written records
The reached end boolean represents that the processor has reached the
end of the log when it tried to read the next record. It could be that
the stream processor wrote new follow-up records to the dispatcher, but
these are not yet available on the log. Since we can know about this,
the stream processor should also be aware of this when determining that
it has reached the end.

Note that reaching the end is thus something different then the log
having a next record. Reaching the end here means that the stream
processor is actually done until a new command is written from outside.

(cherry picked from commit 2141f879adf5efddfab7cc8f2bb324bd14b12990)