fix(engine): consider non-follow-up records
This check would already work in most cases. However, it should also support the case that the last record was not written by this stream processor. In that case the position of the record is higher than the lastWrittenPosition

Co-authored-by: Philipp Ossler <philipp.ossler@gmail.com>
(cherry picked from commit 52f424bdde8fdc845ace73a2ea3956438871ee27)