refactor: enable waiting for manual snapshot
Taking a manual snapshot with `forceSnapshot` was completely async.
Because this has made some tests flaky, we now return a future that
is completed when the snapshot was taken (or skipped).
This allows us to wait for snapshots in tests.

(cherry picked from commit 16d7cd1dc103df73d507acbcf39c063bcdcdae3b)