refactor(engine): align zeebe transformers
The zeebe transformers used slightly different implementation patterns,
which leads to different call patterns. Should the tranformer be called
if the specific extension element is not in the model? I think the
validators should make sure that all model parts are defined as
expected. The transformers should simply be able to deal with the case
that it was optional. This matters, because we're going to change 1 of
the tasks that currently always has a task definition but it will be
optional after the future changes (i.e. business rule task).

Therefore, I wanted to align the implementation of these classes so
their usage can be kept the same over multiple elements that may contain
these extension elements.

Lastly, I chose to move the getSingleExtensionElement call to the caller
side, to make it more clear at the caller side what this transformer
does. This also makes it very easy to move the 'if element is null do
nothing' check into the caller side, if we do happen to decide that the
transformers should only be called when they have the element.