feat(engine): introduce executable business rule task
The business rule task is either a job worker task or it has a called
decision. To represent this, the class should be able to be both: it is
a job worker task and it can have a called decision. In reality only 1
of these should be present. This should at the minimum be checked by the
validators, but we could also consider adding a safety check in this
class. For now, I've left it out.