feat: support task with called decision processing
The business rule task can be either a job worker task or have a called
decision. If it's a job worker task, we can simply reuse the existing
JobWorkerTaskProcessor. If it has a decision to call, we can process it
differently. This requires some specialised behavior classes, because
we can't simply use something like:

```
private final BpmnElementProcessor<ExecutableJobWorkerTask> jobWorkerTaskBehavior;
private final BpmnElementProcessor<ExecutableBusinessRuleTask> calledDecisionBehavior;
```

This would fail because at the calling site of the .onX methods we'll
notice that the types don't match anymore.

Instead, we can simply define a BusinessRuleTaskBehavior. We can implement
it for the job worker task using delegation, while we can implement it
independently for the called decision behavior.