test(engine): verify called decision processing
This adds the failing tests for the called decision behavior when
processing a business rule task that specifies a called decision id.

These tests are inspired by the JobWorkerElementTest and follow a
similar structure, with the major difference that instead of verifying
the creation of a job, it verifies the writing of the result variable.