feat(engine): process business rule task with called decision
This introduces a new BpmnDecisionBehavior class to provide decision
behavior to the bpmn processors, and it is immediately used in the
BusinessRuleTaskProcessor with called decision behavior.

The implementatino is pretty straightforward. It needs to:
 - find the drg in the state (this has some potential failure cases)
 - evaluate the decision in the drg (this can also fail)
 - trigger a process event with the result variable if successfully
   evaluated.

The process event triggering record is used to accomodate the async
nature of evaluating the decision on element activations, while the
result variable is only used on completino of the element. That is, the
output mapping behavior uses it to determine whether it will propagate
it as a process variable, or write it as a local variable and use it in
the output mappings. This combined with the potential to raise incidents
as part of the output mappings behavior, means that the value must be
available in state and thus on record.

The event trigger that this creates is cleaned up by the
ProcessInstanceElementCompletedApplier as the normal part of cleanup
involved with the element being completed.

All failure cases mentioned above result in an incident on the business
rule task during while activating, with the exception of the incidents
raised during output mapping.

Lastly, it felt a bit cleaner to provide a smaller interface to the
BpmnDecisionBehavior.evaluateDecision method. So, I extracted the
ExecutableCalledDecision interface from the ExecutableBusinessRuleTask.