test(engine): verify called decision incident behavior
This adds the failing tests for the called decision behavior related to
incident when processing a business rule task that specifies a called
decision id.

Specifically, it makes sure that failing to evaluate the decision leads
to a resolvable incident. In addition, it makes sure that incidents due
to output mappings behave the same for such business rule tasks as for
service tasks.

Small refactorings of the OutputMappingIncidentTest were necessary to
deploy the dmn along with the process model.