feat(util): add Either.ofOptional convenience method
This method can be used to easily convert an Optional into an Either.

Optional provides an .orElse method to successfully retrieve a result
from it, but this always has to be the same type as the Optional's type
parameter. In the case of Either, we want to be able to express the
'other' value (i.e. left)  with a potentially different type.

Users are already used to using .orElse in the context of Optionals.
When converting an Optional into an Either, it then makes sense to
provide a similarly named method to supply the left value.