build: use windows-2022
Switches virtual environment for windows to windows-2022. The hope is
this newer environment will be marginally faster, which seems to be
anecdotally the case. It's possible it will be less stable as it's quite
new, so we should be prepared to switch back if that's the case.

(cherry picked from commit 1d71e74a962683cadee7d13910722670584a80fd)