build: tune Java for faster startup
Adds specific Java configuration options tuned for faster startup, in
the hopes that the Windows initial maven call will be faster. Turning on
tiered compilation with a max level of 1 is essentially using the C1
compiler (or as it used to be called, the client compiler, optimized for
startup at the cost of runtime performance).