fix: throw instead of overwriting timer
Here we throw an exception when a timerId was generated that already
exists in the timerJobMap.
Throwing here will fail when scheduling a new timer or when resubmitting
a recurring timer.
This should never happen, but we have one heapdump where it looks like
it did: https://github.com/camunda-cloud/zeebe/issues/8776