test(qa): verify authority overriding
Adds a test case to the SecurityTest that verifies that the authority
(used in the TLS handshake) can be overridden.

A minor refactoring was necessary to make it possible to create the
client in each test, while also passing the same configuration to the
cluster rule. The client rule is not needed anymore because of this.