refactor(logstreams): introduce canWriteAdditionalEvent to batch writer
Replaces the API to get the batch length with an additional event with a
simpler predicate method, `canWriteAdditionalEvent(int)`. This avoids
breaking the abstractions and having the writer know how the dispatcher
will compute the framed and aligned batch length, and instead simply
delegates deciding whether the batch, with the additional event of the
given length, can be written to the dispatcher or not.