feat(protocol): add CALLED_DECISION_ERROR
This error type will be used when a decision is referenced that cannot
be found, e.g. it is used in the called decision element of a business
rule task, while no dmn model containing a decision with that id was
deployed.

This change does not affect the other ErrorTypes. I've tested creating
multiple incidents of different types with code from before this commit,
then shutting down the engine and restarting it with the new ErrorType
added. All incidents could be read from the log with the correct type
and it was possible to resolve all of them correctly.

However, it might make sense to add short values similar to our other
Enums in the future, to guarantee forwards compatibility.