test(engine): verify called element error incident
Adds a failing test case to make sure that an incident is correctly
created if decision is not deployed. It refers to `unknown_decision_id`
to be independent of the other test cases, as it's the only one to not
deploy a dmn resource. If we would've used the same `jedi-or-sith`
decision id, then the order of test case execution would matter.