feat: config to disable checks in `ColumnFamily`
When we will switch to using methods that check preconditions, we need a
config to disable the check. This is important in case this comes with
an unexpected performance impact, or more importantly, to disable the
checks in case we used a wrong methods somewhere (e.g. insert instead of
upsert)