feat(dispatcher): add new #canClaimFragmentBatch API
Adds a new public API method to the Dispatcher,
`#canClaimFragmentBatch(int, int)`. This method determines whether the batch
with the given fragment count and of the given length can actually be
claimed for this dispatcher instance.

This method is mostly useful to determine if something will be claimable
before you wish to claim it, mostly to avoid having to abort your claim
unnecessarily if the batch isn't finished yet when calling this method.

(cherry picked from commit 9281e8126a18d4b1dfbaeb925be569af9334571e)