feat(logstreams): extend batch writer with write size probe
Introduces a new API to the `LogStreamBatchWriter` API,
`#canWriteAdditionalEvent(int)`. This API lets one probe whether a batch
including an additional event of the given size could be written.

While this could be done via tryWrite, there are cases, such as with the
job activation, where you want to know if adding data to a record (not
to the writer, but the record) will cause it to be unwritable, at which
point you want to handle it differently. Rolling back changes to the
record itself is more complicated than to the writer, so as a compromise
you can probe the writer to know if it will be able to write more bytes
before writing them, even, as in our example, to your DTO.

(cherry picked from commit e706d940375cefd768848157ddf595c6aabe40f3)