refactor(engine): extend TypedEventWriter to deal with sizes
Adds size probing capabilities to the `TypedEventWriter`, namely one
`#canWriteEventOfLength(int)` which checks if an additional event of the
given length could be written. This simply delgates to the batch writer
under the hood the normal case, with some sane overrides for other
implementations.

Additionally delegates to the writer to report/determine the
`maxEventLength` instead of getting it from the `ProcessingContext`.

The reasoning for both is that these are writer specific properties:
it's the writer's responsibility to know what it can write, and whether
it can write it or not.

(cherry picked from commit 3d8af7be77c0db28c60a1608fbd6d6c3f3a8863d)