test(util): remove util's EitherAssert
Removes the customized, generated `EitherAssert` in the util module. This was
only used for testing `Either` itself, but there was only marginal value
to it. Furthermore, we already have another `EitherAssert` available in
the `test-util` module, which is the one meant to be reused in all the
other modules as a test utility.