test: use engine rule per test case
We want to add a test case that requires the decision model to not be
deployed, so an incident is created which can be resolved by deploying
the model. However, some of the other test cases also require the model
to be deployed. So the order of execution matters.

We can circumvent this by switching the engine rule from a class rule to
a regular test case rule. This means we get a fresh engine per test
case.