feat(protocol): add DECISION_EVALUATION_ERROR
This error type will be used when an attempt to evaluate a deployed
decision has failed. E.g. a business rule task with called decision is
activating, but one of the variables required to evaluate the referred
decision is missing.

This error type differs from the CALLED_DECISION_ERROR, because it helps
users to understand better what they need to do resolve the problem. In
that case, a decision was referred that is unknown (e.g. no dmn resource
containing a decision with that id has been deployed yet). That is
different from this case, where the decision could be found, but could
not be correctly evaluated. Details on the reason are generally provided
along the error type to help the user to resolve the underlying problem.

This change does not affect the other ErrorTypes. I've tested creating
multiple incidents of different types with code from before this commit,
then shutting down the engine and restarting it with the new ErrorType
added. All incidents could be read from the log with the correct type
and it was possible to resolve all of them correctly.

However, it might make sense to add short values similar to our other
Enums in the future, to guarantee forwards compatibility.