test(engine): verify evaluation of decisionId expression
A CalledDecision with an expression as its decisionId should be able to
call the decision based on the evaluation result of the expression.

For example, if the decisionId is an expression `= decisionId`, then the
value of the `decisionId` variable should be used as the id of the
decision that is called.

The test can verify that it works by asserting that a result variable
has been created for the process instance. Note that no decision with id
`=decisionIdVariable` exists, so the test fails until the evaluation
result of the expression is used, instead of the static value.