test(engine): raise incident for decision id expression
An incident should be raised when the decision id expression fails to be
evalutated successfully. The error type should be EXTRACT_VALUE_ERROR
like the other expression evaluation failures.

In addition, it should be possible to resolve the incident by removing
the problem (i.e. set the missing variable in this case).