feat: transform decisionId to expression
The decisionId is validated to be valid expression (or static value) by
the ZeebeRuntimeValidators. So we can safely transform it to an
expression by parsing it using the expression language.

To keep the behavior the same, the BpmnDecisionBehavior needs to extract
the original string representation of the expression. For static values,
this is simply the static value again. For expressions it would be the
entire expression in string representation.

Since expressions are not yet evaluated, this keeps the logic the same
from a user's perspective. But the decisionId is now additionally parsed
to allow evaluation to be added later.