test(zbctl): properly detect command timeout in tests
As shown in [this issue](https://github.com/golang/go/issues/21880),
running a command with a deadline or timeout causes the process to be
killed properly, but the error returned is that it was killed, without
any more information. To know that it timed out, one has to use the
context itself via `context.Err()`. We now do this and also set the
timeout flag of the command to add a second, to ensure we can properly
detect time outs, as otherwise it's difficult to know why the command
exited since it always uses the same exit code on error.