docs: add javadoc
Iterating over ArrayValue is not thread-safe. Iterating over it will modify the buffer, doing this from multiple threads at once will result in exceptions. This is unexpected behaviour which is not documented properly. To warn others and hopefully save some time debugging in the future I have added some javadoc to this method.