refactor(protocol): generate immutable values in protocol
Moves the generation of the immutable protocol values from
`protocol-jackson` to `protocol`. This has the advantage of providing
easy to use immutable implementations, especially useful for testing,
but potentially as well for easy copying, transforming mutable records
into immutable values via deep-copy.

Renames the `clone` method to `copyOf`. This lets us define it as a
`@Value.NonAttribute` method with a default implementation.
Additionally, this was flagged as warning as the clone method is a
special method in Java, and our interfaces were not following its
semantics, nor implementing the `Cloneable` interface.

The `protocol-jackson` module now only provides a `ZeebeProtocolModule`,
which must be added to an `ObjectMapper`. Doing so will let the mapper
serialize and deserialize any of the abstract protocol value, without
even requiring the user to know about the `Immutable*` variants.