build: reduce BOM to only public modules
Reduces the set of modules specified in the BOM to only these modules
which are meant for third-party consumption. This means, only the
modules which are expected and prepared to be used in projects outside
of Zeebe. All other so called internal modules are defined in the
parent.

The goal here is to make it more explicit to third party users which
modules they can safely include in their projects, and which are more
likely to change/disappear/break over time, and thus they should not
rely on. They can still import them, but they must now do so explicitly
and at their own risk.