fix: notify new `SnapshotReplicationListener`s about ongoing replication
We are using `SnapshotReplicationListener`s to transition to inactive
when a snapshot replication starts. Snapshot replication can start
before any listeners have registered, which means that the listener will
only be notified about snapshot replication finishing, triggering a
transition to follower without first transitioning to inactive.

Here we are keeping track of ongoing snapshot replication to immediately
notify listeners when they are registering.