refactor: transform class into record
The VariableValue class is only used in variable mapping tests and
exists only as an easy way to create and compare variable values. This
makes for an ideal record candidate.

Note that equals method should stay overridden, as the value should not
be compared with string equality, but with json string equality (e.g. an
object with 2 properties can have either property before the other in
the string representation while they are still equal, child order
doesn't matter).