test(engine): variable mappings support date values
This adds failing tests, that can be fixed in subsequent commits. The
tests verify that both input and output mappings support values that have
a temporal type. It assumes that the FEEL-function to create a date,
time, datetime and duration object is availeble, e.g. `date(string)`.

I considered adding these tests in the expression-language, but the
expression language is already able to evaluate expression resulting in
a value with temporal types. What is still missing is the transformation
into a variable value. No tests for this exists in the
expression-language. We could add tests there if we have additional edge
cases to consider. If not (and I can't think of any right now), we
should simply test the usage of this, so: input and output mappings that
alter variables on a process instance.