feat(feel): transform temporal feel values to msgpack
Feel expressions can result in a value of a temporal type (e.g. a date
or duration). When that happens, we want to be able to store that value
in a local process variable. To support that, the feel value must be
transformed into a message pack value. This adds support for that by
adding the necessary case patterns.

Temporal values should be stored in a variable using their string
representation. For the ValTime, the internal value is a ZonedTime which
implements a format method to get this String representation. For the
other temporal Val types, this String can be retrieved using .toString.
Since ValString can already be written, we re-use this to write the new
string representations.

A previous commit added tests for these types.