feat(engine): allow static input mappings
Before input mappings always had to be expressions. It is not clear
anymore why we enforced this, but Zeebe already supports static
expressions based on the value of the field.

This change remove the validation for non static expression on input
mappings to improve the user expierence we can offer in the connectors
use case.

https://github.com/camunda/cloud-connectors/issues/13

The static value will be escaped as a java string, and not as a JSON
string as it would otherwise additionally escape `/` characters which
will be used for URL parameters and additionally escaping them decreases
the user experience.

The current implementation might multiple times escape characters like
`"` and `\`, as they are escaped initially to be given to the FEEL
engine and later when retrieving the value from the variable record.