fix(protocol-jackson): map only direct implementations
Fixes mapping only the direct implementations of a concrete type. Some
of the abstract types extend each other, which result in multiple
possible concrete types. However, in our design, only one concrete type
really ever makes sense. For example, we have ProcessMetadataValue and
Process. Process is a record value which extends ProcessMetadataValue,
whereas ProcessMetadataValue is just an abstract type and has no value
type associated to it. So for records where we map the value type to
ValueType.PROCESS, then we want to use the Process abstract type and the
ImmutableProcess concrete type, as the other option (e.g.
ImmutableProcessMetadataValue) just doesn't make sense. And it's
essentially the same the other way around.