refactor: directly write string value
The temporal values are written as strings, which was already defined by
the ValString case. Instead of creating a new objet `ValString` and
reusing the pattern match, we could directly write a string value by
extracting that into a separate method. This should make it a bit easier
to understand.