build: reduce spotbugs configuration
The feel module contains a `spotbugs/spotbugs-exclude.xml` file that
specifies that the scala classes should be ignored. Thus far, it was
unclear how this file was picked-up by the spotless-maven-plugin,
considering that `parent/pom.xml` specifies the file as
`spotbugs/spotbugs-exclude.xml`. Note that this does not include the
`${project.basedir}` and thus refers to the root project, instead of the
feel module's own file.

The only other module that defines its own spotbugs-exclude file is the
expression-language. The `expression-language/pom.xml` file uses the
`${project.basedir}` to refer to its own file and overwrite the base one
(defined by the build-tools module). This makes it additionally strange
that the exclusion file in the feel module could be picked up by
spotbugs. So what is going on?

Running mvn clean install -Pspotbugs -X shows us:
[DEBUG]   Adding Exclude Filter File
[DEBUG] resource is spotbugs/spotbugs-exclude.xml
[DEBUG] artifact is spotbugs-exclude.xml
[DEBUG] The resource 'spotbugs/spotbugs-exclude.xml' was found as
'/Users/korthout/dev/camunda-cloud/zeebe/feel/spotbugs/spotbugs-exclude.xml'
[DEBUG] location of resourceFile file is
/Users/korthout/dev/camunda-cloud/zeebe/feel/target/spotbugs-exclude.xml

So leaving out the `${project.basedir}` works, but why? Well, it seems
that spotbugs resolves the configuration files using the location of the
pom file. Specifically it resolves them relatively to the POM's parent
directory. We ran into a problem with that while adding the flatten
plugin, because it was changing the POM's location for any plugins that
ran after the flatten plugin. A note has been added to that plugin's
configuration to not make that mistake again.

Coming back to our spotbugs configuration, we can now improve things in
multiple ways:
- remove the property in expression-lanaguage that overwrites the
  spotbugs configuration file's location
- add that same property to the feel-integration module
- change the spotbugs configuration by providing a file to each leaf
  project and completely remove the common configuration

I've chosen to remove the property from the expression-language that
overwrites the spotbugs configuration file's location. My main reason is
that it is not necessary to configure this. Since it is not necessary it
can (and has) lead to inconsistent configuration (see feel and
expression-language difference). And this might happen again when we try
to copy such config. By removing the property, the only configuration
that is left, are the actual spotbugs-exclude.xml files. Copying it to
another module, will make that module use it as well.

(cherry picked from commit 8c33dfa17aa85828be7580ffe764e829eb3d46ea)