build(bom): overwrite wrongly inheritted values
The BOM inherits from the camunda-release-parent, which sets a bunch of
values that don't make sense for zeebe's bom. This overwrites does
values with better fitting alternatives.

Alternatively, this could have inlines the inheritted parent pom. But
this removes central control over the deployment of maven modules.
Perhaps we'll reconsider this later, but for now, it's easier to simply
overwrite the values.

Note that the description of the BOM has been emptied, because zeebe's
parent-pom inherits from the zeebe-bom. This avoids that any submodules
that do not define a description themselves inherit the
camunda-release-parent's description. In the future we may change this
by properly providing a description to each of the poms.